training/requirements.txt
# Training requirements
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0
matplotlib>=3.7.0
